pyproject.toml
[tool.poetry]
name = "mot-meme-generator"
version = "0.1.0"
description = "Motivational meme generator"
authors = ["Matěj Dvořáček"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
pillow = "^10.4.0"
black = "^24.8.0"
isort = "^5.13.2"
pydocstyle = "^6.3.0"
python-docx = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
